refactor(tools): move pdf merge logic
