+++ utils/pdf/src/index.ts
@@ -1,2 +1,1 @@
 export * from './remove-pdf-owner-password'
-export * from './merge-pdf'
