--- utils/pdf/src/merge-pdf/index.ts
@@ -1,1 +1,0 @@
-export { mergePDFs } from './merge-pdf-with-worker'
