--- utils/pdf/src/merge-pdf/merge-pdf.worker.ts
@@ -1,5 +1,0 @@
-import { mergePDFs } from './merge-pdf'
-
-onmessage = async (e: MessageEvent<File[]>) => {
-  postMessage(await mergePDFs(e.data))
-}
